+++ dascore/proc/basic.py
@@ -198,48 +198,6 @@ def standardize(
 
 
 @patch_function()
-def integrate(patch: PatchType, dim: str, keep_dims=False) -> PatchType:
-    """
-    Integrate along a specified dimension using simpson's composite rule.
-
-    Parameters
-    ----------
-    patch
-        Patch object for integration.
-    dim
-        The dimension along which to integrate.
-    keep_dims
-        If True, collapse (remove) the integration dimension.
-
-    Examples
-    --------
-    >>> import dascore as dc
-    >>> patch = dc.get_example_patch()
-    >>> integrated = patch.integrate(dim='time')
-    """
-
-
-@patch_function()
-def diff(patch: PatchType, dim: str) -> PatchType:
-    """
-    Differentiate along specified dimension.
-
-    Parameters
-    ----------
-    patch
-        The patch to differentiate
-    dim
-        The dimension along which to differentiate.
-
-    Examples
-    --------
-    >>> import dascore as dc
-    >>> patch = dc.get_example_patch()
-    >>> integrated = patch.diff(dim='time')
-    """
-
-
-@patch_function()
 def apply_operator(patch: PatchType, other, operator) -> PatchType:
     """
     Apply a ufunc-type operator to a patch.
