+++ tests/test_proc/test_differentiate.py
@@ -1,0 +1,3 @@
+"""
+Module for testing differentiation of patches.
+"""
